refactor(core): move DataProvider interface to factory
