+++ packages/botonic-core/src/data-provider/index.ts
@@ -1,4 +1,3 @@
-export * from './data-provider'
 export * from './dynamodb-data-provider'
 export * from './dynamodb-utils'
 export * from './factory'
