+++ packages/botonic-core/src/data-provider/local-data-provider.ts
@@ -2,7 +2,7 @@ import { JsonDB } from 'node-json-db'
 import { Config } from 'node-json-db/dist/lib/JsonDBConfig'
 
 import { BotonicEvent, User } from '../models'
-import { DataProvider } from './data-provider'
+import { DataProvider } from './factory'
 
 export class LocalDevDataProvider implements DataProvider {
   private readonly DB_PATH = 'tmp/localDb'
